broken down components much more granularly
