+++ src/lib/Burger.svelte
@@ -1,0 +1,32 @@
+<script>
+  let showMobileNavMenu = "hidden"
+  function toggleShowMobileNavMenu() {
+    if (showMobileNavMenu === "hidden") {
+      showMobileNavMenu = ""
+      return
+    }
+    showMobileNavMenu = "hidden"
+  }
+</script>
+
+<div aria-pressed="false"
+     class="fixed top-0 right-0 z-50 m-5 origin-right md:hidden"
+     id="burger"
+     role="button">
+  <a href="#burger" on:click={toggleShowMobileNavMenu}>
+    <div class="space-y-2">
+      <div class="w-8 bg-slate-300 h-0.5"></div>
+      <div class="w-8 bg-slate-300 h-0.5"></div>
+      <div class="w-8 bg-slate-300 h-0.5"></div>
+    </div>
+  </a>
+</div>
+<!-- nav popup thing -->
+<div class="fixed top-16 right-5 w-36 to-slate-950 bg-gradient-to-r from-50% from-slate-900 rounded-md shadow-lg shadow-black md:hidden {showMobileNavMenu}">
+  <ul class="m-5 flex flex-col gap-2 text-xl tracking-wide">
+    <li><a href="#splash" on:click={toggleShowMobileNavMenu}>Home</a></li>
+    <li><a href="#about" on:click={toggleShowMobileNavMenu}>About</a></li>
+    <li><a href="#work" on:click={toggleShowMobileNavMenu}>Work</a></li>
+    <li><a href="#projects" on:click={toggleShowMobileNavMenu}>Projects</a></li>
+  </ul>
+</div>
