--- src/lib/Container.svelte
@@ -1,3 +1,0 @@
-<div class="grid grid-cols-1 text-slate-50 md:grid-cols-2 xl:grid-cols-3" id="container">
-  <slot></slot>
-</div>
